requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-rest-swagger==2.1.2
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
openapi-codec==1.3.2
pep8==1.7.1
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pytest==3.4.2
python-decouple==3.1
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
